Fix incorrect recursion in random_mudpie_sample
The recursive function call was missing an argument, so sometimes a framerate of 1 was passed and since count has a default assignment of zero the recursive call didn't throw an immediate error with only three arguments. This would occasionally result in a one-frame sample returning to the main logic. This resulted in fatal errors during the filtering and declicking stages.