Update mudpie_sample_generator.py
Define smoothing function with hanning window to smooth the randomly generated amplitude envelope.

Also set beginning and ending of envelope to 0 before smoothing as the interpolated curves can be a bit off at the ends.

Clip the envelope to ensure values stay between 0 and 1.

Finish function that applies envelope to sample.