Update mudpie_sample_generator.py
DC offset wasn't due to faulty bandpass or highpass but was due to signal translation that occured when writing the wav files. Changed save parameters to rectify this and removed the second highpass step I was testing.

Tweaked gate parameters again for smoother sound.

Commented out saving the raw_sample as I feel this script is now production ready enough.